+++ NEWS.md
@@ -1,4 +1,5 @@
 # SDMtune 1.2.0.9000
+* `plotCor()`gains the argument `text_size` to tweak the size of the text in the plot
 
 # SDMtune 1.2.0
 * `SDMtune` migrated to package `terra` to handle spatial data. Objects created with the package `raster` will not be supported in future releases